apps/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
itsdangerous==2.1.2

# HTTP Client (for OAuth validation)
httpx==0.26.0

# Object Storage (MinIO/S3)
boto3==1.34.34

# Azure Speech SDK
azure-cognitiveservices-speech==1.34.0

# Admin Panel
sqladmin==0.16.0

# Redis
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.1.0

# Monitoring
prometheus-client==0.19.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-cors==1.0.0
